esp32-build: stl containers to use allocator which prefers SPI-/PSRAM
This change is to implement and use a std::allocator<> which prefers
SPI-connected PSRAM when available, so the fragment data is stored in
the much larger PSRAM from the outset (rather than using available DRAM
first).
This fixes issues eg. during encoding which can otherwise be caused for
other tasks/processes (which may require DRAM) when a large number of
data fragments are being decoded (ie. when a large payload has been
encoded into a large number of fragments).